feat: #1 separate vocal line, add lyrics
This change breaks out the melody line into its own variable that is
reference later, making things a bit easier to read. I also went ahead
and added all of the stanzas of the lyrics. May add the 5th stanza as a
standalone block at the end, but we shall see once I get the
accompaniment part engraved.